# apps/api/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pickle Go API
  description: |
    Pickle Go is a web-first pickleball event management platform.
    This API provides endpoints for user authentication, event management, and registration.
  version: 1.0.0
  contact:
    name: Pickle Go Team
    url: https://picklego.tw
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.picklego.tw/api/v1
    description: Production server
...
paths:
...
  /events:
    get:
      tags:
        - Events
      summary: List Events
      description: List events with optional filtering by location and skill level
      operationId: listEvents
      parameters:
        - name: lat
          in: query
          description: Latitude for geo-based search
          schema:
            type: number
            format: double
        - name: lng
          in: query
          description: Longitude for geo-based search
          schema:
            type: number
            format: double
        - name: radius
          in: query
          description: Search radius in meters (default 10000)
          schema:
            type: integer
            default: 10000
            maximum: 50000
        - name: skill_level
          in: query
          description: Filter by skill level
          schema:
            $ref: '#/components/schemas/SkillLevel'
        - name: status
          in: query
          description: Filter by event status
          schema:
            $ref: '#/components/schemas/EventStatus'
        - name: limit
          in: query
          description: Number of results to return
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventListResponse'

    post:
      tags:
        - Events
      summary: Create Event
      description: Create a new event
      operationId: createEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreateEventResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{id}:
    get:
      tags:
        - Events
      summary: Get Event
      description: Get event details by ID
      operationId: getEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Events
      summary: Update Event
      description: Update an existing event (host only)
      operationId: updateEvent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '403':
          description: Not the event host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Events
      summary: Cancel Event
      description: Cancel an event (host only)
      operationId: deleteEvent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '403':
          description: Not the event host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        email:
          type: string
          format: email
    SkillLevel:
      type: string
      enum:
        - beginner
        - intermediate
        - advanced
        - expert
        - any
    EventStatus:
      type: string
      enum:
        - open
        - full
        - cancelled
        - completed
    Location:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        google_place_id:
          type: string
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        host:
          $ref: '#/components/schemas/User'
        title:
          type: string
        description:
          type: string
        event_date:
          type: string
          format: date
        start_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        end_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        location:
          $ref: '#/components/schemas/Location'
        capacity:
          type: integer
          minimum: 4
          maximum: 20
        confirmed_count:
          type: integer
        waitlist_count:
          type: integer
        skill_level:
          $ref: '#/components/schemas/SkillLevel'
        skill_level_label:
          type: string
        fee:
          type: integer
          minimum: 0
          maximum: 9999
        status:
          $ref: '#/components/schemas/EventStatus'
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        total:
          type: integer
        has_more:
          type: boolean
    CreateEventRequest:
      type: object
      required:
        - event_date
        - start_time
        - location
        - capacity
        - skill_level
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        event_date:
          type: string
          format: date
        start_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        end_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        location:
          type: object
          required:
            - name
            - lat
            - lng
          properties:
            name:
              type: string
            address:
              type: string
            lat:
              type: number
              format: double
            lng:
              type: number
              format: double
            google_place_id:
              type: string
        capacity:
          type: integer
          minimum: 4
          maximum: 20
        skill_level:
          $ref: '#/components/schemas/SkillLevel'
        fee:
          type: integer
          minimum: 0
          maximum: 9999
          default: 0
    CreateEventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        share_url:
          type: string
          format: uri
    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        event_date:
          type: string
          format: date
        start_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        end_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        capacity:
          type: integer
          minimum: 4
          maximum: 20
        skill_level:
          $ref: '#/components/schemas/SkillLevel'
        fee:
          type: integer
          minimum: 0
          maximum: 9999
        status:
          type: string
          enum:
            - open
            - full
            - cancelled
...